refactor: several changes base on pr comments
